requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
cffi==1.14.0
cryptography==2.9
dj-database-url==0.5.0
Django==3.0.5
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
